Refactor risk engine
