+++ programs/marginfi/src/state/mod.rs
@@ -2,4 +2,5 @@ pub mod fee_state;
 pub mod marginfi_account;
 pub mod marginfi_group;
 pub mod price;
+pub mod risk_engine;
 pub mod staked_settings;
